+++ Solution/RiskOfChaos/EffectDefinitions/World/Gravity/RotateGravity.cs
@@ -18,19 +18,33 @@ namespace RiskOfChaos.EffectDefinitions.World.Gravity
     [EffectConfigBackwardsCompatibility("Effect: Random Gravity Direction (Lasts 1 stage)")]
     public sealed class RotateGravity : NetworkBehaviour
     {
-        const float MAX_DEVITATION_MIN_VALUE = 0f;
-        const float MAX_DEVITATION_MAX_VALUE = 70f;
+        const float DEVITATION_MIN_VALUE = 0f;
+        const float DEVITATION_MAX_VALUE = 90f;
 
         [EffectConfig]
-        static readonly ConfigHolder<float> _maxDeviation =
-            ConfigFactory<float>.CreateConfig("Max Rotation Angle", 30f)
-                                .Description("The maximum amount of deviation (in degrees) that can be applied to the gravity direction")
-                                .AcceptableValues(new AcceptableValueRange<float>(MAX_DEVITATION_MIN_VALUE, MAX_DEVITATION_MAX_VALUE))
+        static readonly ConfigHolder<float> _minDeviation =
+            ConfigFactory<float>.CreateConfig("Min Rotation Angle", 25f)
+                                .Description("The minimum amount of deviation (in degrees) that can be applied to the gravity direction")
+                                .AcceptableValues(new AcceptableValueRange<float>(DEVITATION_MIN_VALUE, DEVITATION_MAX_VALUE))
                                 .OptionConfig(new StepSliderConfig
                                 {
                                     FormatString = "{0:F1}",
-                                    min = MAX_DEVITATION_MIN_VALUE,
-                                    max = MAX_DEVITATION_MAX_VALUE,
+                                    min = DEVITATION_MIN_VALUE,
+                                    max = DEVITATION_MAX_VALUE,
+                                    increment = 0.5f
+                                })
+                                .Build();
+
+        [EffectConfig]
+        static readonly ConfigHolder<float> _maxDeviation =
+            ConfigFactory<float>.CreateConfig("Max Rotation Angle", 50f)
+                                .Description("The maximum amount of deviation (in degrees) that can be applied to the gravity direction")
+                                .AcceptableValues(new AcceptableValueRange<float>(DEVITATION_MIN_VALUE, DEVITATION_MAX_VALUE))
+                                .OptionConfig(new StepSliderConfig
+                                {
+                                    FormatString = "{0:F1}",
+                                    min = DEVITATION_MIN_VALUE,
+                                    max = DEVITATION_MAX_VALUE,
                                     increment = 0.5f
                                 })
                                 .Build();
@@ -53,7 +67,9 @@ namespace RiskOfChaos.EffectDefinitions.World.Gravity
             Xoroshiro128Plus rng = new Xoroshiro128Plus(_effectComponent.Rng.nextUlong);
 
             float maxDeviation = _maxDeviation.Value;
-            _gravityRotation = QuaternionUtils.Spread(Vector3.down, maxDeviation / 3f, maxDeviation, rng);
+            float minDeviation = Mathf.Min(_minDeviation.Value, maxDeviation);
+
+            _gravityRotation = QuaternionUtils.Spread(Vector3.down, minDeviation, maxDeviation, rng);
 
             Log.Debug($"Gravity angle: {Vector3.Angle(Vector3.down, _gravityRotation * Vector3.down)}");
         }